+++ src/CHANGELOG.md
@@ -22,6 +22,7 @@
 - Utils: Add spaCy's Azerbaijani word tokenizer and stop word list
 - Utils: Add spaCy's Greek (Ancient) word tokenizer and stop word list
 - Utils: Add ssg's Thai syllable tokenizer
+- Work Area: Add Overview - Count of Syllables / Token Length in Syllable / Type Length in Syllable / Syllable Length in Character
 - Work Area: Add Work Area - Search Settings - Multi-search Mode - Insert
 
 ### ✨ Improvements